requirements.txt
streamlit
groq
sentence-transformers
faiss-cpu
pypdf
scikit-learn
numpy
python-dotenv
nicegui
